+++ sdk/main/src/helper/DurationConverter.cc
@@ -30,7 +30,7 @@ std::chrono::seconds fromProtobuf(const proto::Duration& duration)
 }
 
 //-----
-proto::Duration* toProtobuf(const std::chrono::duration<double>& duration)
+proto::Duration* toProtobuf(const std::chrono::duration<int64_t>& duration)
 {
   auto proto = new proto::Duration;
   proto->set_seconds(std::chrono::duration_cast<std::chrono::seconds>(duration).count());
